+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes will be documented in this file.
 
 ### Changed
 
+* `Parameter` now includes the `account`, `exchange`, `symbol` tuple
 * `ParametersUpdate` was renamed (from `ParameterUpdate`)
 * `TraceInfo` now default initializing to current time
 * `clock` promoted (from core library)